\name{copyNFIboundary_spatial}
\alias{copyNFIboundary_spatial}
\title{copyNFIboundary spatial}
\description{Copy the administrative-boundary sidecar between objects.}
\usage{copyNFIboundary_spatial(from, 
    to)}
\arguments{
  \item{from}{Source object carrying an administrative-boundary sidecar.}
  \item{to}{Target object that will receive the boundary sidecar.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





