\name{copyNFIgeometry_spatial}
\alias{copyNFIgeometry_spatial}
\title{copyNFIgeometry spatial}
\description{Copy the plot-geometry sidecar between NFI objects.}
\usage{copyNFIgeometry_spatial(from, 
    to)}
\arguments{
  \item{from}{Source object carrying plot-geometry and related spatial attributes.}
  \item{to}{Target object that will receive the spatial attributes.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





