\name{copyNFIspatial_sidecars}
\alias{copyNFIspatial_sidecars}
\title{copyNFIspatial sidecars}
\description{Copy both plot-geometry and administrative-boundary sidecars.}
\usage{copyNFIspatial_sidecars(from, 
    to)}
\arguments{
  \item{from}{Source object carrying plot-geometry and administrative-boundary sidecars.}
  \item{to}{Target object that will receive both spatial sidecars.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





