\name{getNFIboundary_spatial}
\alias{getNFIboundary_spatial}
\title{getNFIboundary spatial}
\description{Return the optional administrative-boundary sidecar.}
\usage{getNFIboundary_spatial(x)}
\arguments{
  \item{x}{Object from which to retrieve the administrative-boundary sidecar stored in the \code{nfi_boundary} attribute.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





