\name{getNFIgeometry_spatial}
\alias{getNFIgeometry_spatial}
\title{getNFIgeometry spatial}
\description{Return the plot-geometry sidecar stored in a spatial NFI object.}
\usage{getNFIgeometry_spatial(x)}
\arguments{
  \item{x}{Object from which to retrieve the plot-geometry sidecar stored in the \code{nfi_geometry_registry} attribute..}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





