\name{hasNFIboundary_spatial}
\alias{hasNFIboundary_spatial}
\title{hasNFIboundary spatial}
\description{Test whether an object contains a valid boundary sidecar.}
\usage{hasNFIboundary_spatial(x)}
\arguments{
  \item{x}{Object to test for a valid administrative-boundary sidecar.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





