\name{hasNFIgeometry_spatial}
\alias{hasNFIgeometry_spatial}
\title{hasNFIgeometry spatial}
\description{Test whether an object contains a valid plot-geometry sidecar.}
\usage{hasNFIgeometry_spatial(x)}
\arguments{
  \item{x}{Object to test for a valid, non-empty plot-geometry sidecar.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





