\name{plot.readNFI_spatial}
\alias{plot.readNFI_spatial}
\title{plot readNFI spatial}
\description{Plot a spatial \code{\link{readNFI}} object.}
\usage{\method{plot}{readNFI_spatial}(x, 
    y = NULL, ...)}
\arguments{
  \item{x}{A spatial \code{readNFI_spatial} object to plot.}
  \item{y}{Optional character vector of variable names or numeric vector of column positions to plot; \code{NULL} uses the default numeric variables.}
  \item{\dots}{Additional arguments passed to the internal spatial plotting engine, including options such as \code{vars}, \code{boundary}, and \code{engine}.}
}



\author{Wilson Lara [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>),
Cristobal Ordonez [aut] (ORCID:
<https://orcid.org/0000-0001-5354-3760>),
Aitor Vázquez-Veloso [aut] (ORCID:
<https://orcid.org/0000-0003-0227-506X>),
Felipe Bravo [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





