\name{Rbasifor-package}
\alias{Rbasifor-package}
\alias{Rbasifor}
\docType{package}
\title{Retrieval and Processing of Data of the Spanish National Forest Inventory}

\description{Data sets of the Spanish National Forest Inventory (NFI) are processed computing dendrometric statistics. Function dendroMetrics() controls most of the in-package routines.}

\details{

The DESCRIPTION file:
\packageDESCRIPTION{Rbasifor}
\packageIndices{Rbasifor}
Maintainer: \packageMaintainer{Rbasifor}

}
\author{Wilson Lara <wilarhen@gmail.com>, Felipe Bravo <fbravo@pvs.uva.es>, Cristobal Ordonez <a_cristo@pvs.uva.es>}

\keyword{ package }


