\name{externalMetrics}
\alias{externalMetrics}
\title{externalMetrics}

\usage{externalMetrics(x, var = c("d", 
    "h", "ba", "n", "Hd"), 
    levels = NULL, design, 
    colmap = list(d = c("d", 
        "dbh", "diameter", 
        "diameter_mm"), 
        h = c("h", "height", 
            "height_m")), 
    d_unit = c("mm", 
        "cm")[1], h_unit = c("m", 
        "dm", "cm")[1], 
    keep_cols = NULL, 
    domheight_fun = get0("domheight", 
        mode = "function", 
        inherits = TRUE), 
    schema = NULL)}
\arguments{
  \item{x}{x }
  \item{var}{var }
  \item{levels}{levels }
  \item{design}{design }
  \item{colmap}{colmap }
  \item{d_unit}{d_unit }
  \item{h_unit}{h_unit }
  \item{keep_cols}{keep_cols }
  \item{domheight_fun}{domheight_fun }
  \item{schema}{schema }
}



\author{Wilson Lara <wilarhen@gmail.com> [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>), Cristobal Ordonez <angelcristobal.ordonez@uva.es> [aut] (ORCID: <https://orcid.org/0000-0001-5354-3760>), Aitor Vázquez-Veloso <aitor.vazquez.veloso@uva.es> [aut] (ORCID: <https://orcid.org/0000-0003-0227-506X>), Felipe Bravo <felipe.bravo@uva.es> [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





