\name{print.concentric_design}
\alias{print.concentric_design}
\title{Print a concentric plot design}
\description{Display the main components of a \code{"concentric_design"} object:
design name, subplot radii, minimum diameters, and expansion factors.}
\usage{\method{print}{concentric_design}(x, 
    ...)}
\arguments{
  \item{x}{Object of class \code{"concentric_design"}.}
  \item{\dots}{Further arguments passed to methods. Currently unused.}
}



\author{Wilson Lara <wilarhen@gmail.com> [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>), Cristobal Ordonez <angelcristobal.ordonez@uva.es> [aut] (ORCID: <https://orcid.org/0000-0001-5354-3760>), Aitor Vázquez-Veloso <aitor.vazquez.veloso@uva.es> [aut] (ORCID: <https://orcid.org/0000-0003-0227-506X>), Felipe Bravo <felipe.bravo@uva.es> [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





