\name{print.inventory_design}
\alias{print.inventory_design}
\title{Print a generic inventory design}
\description{Display the main components of an \code{"inventory_design"} object:
design name, diameter thresholds, sampled areas, and expansion
factors.}
\usage{\method{print}{inventory_design}(x, 
    ...)}
\arguments{
  \item{x}{Object of class \code{"inventory_design"}.}
  \item{\dots}{Further arguments passed to methods. Currently unused.}
}



\author{Wilson Lara <wilarhen@gmail.com> [aut, cre] (ORCID: <https://orcid.org/0000-0003-3527-1380>), Cristobal Ordonez <angelcristobal.ordonez@uva.es> [aut] (ORCID: <https://orcid.org/0000-0001-5354-3760>), Aitor Vázquez-Veloso <aitor.vazquez.veloso@uva.es> [aut] (ORCID: <https://orcid.org/0000-0003-0227-506X>), Felipe Bravo <felipe.bravo@uva.es> [aut] (ORCID: <https://orcid.org/0000-0001-7348-6695>)}





